test: add tests to all modules
Added tests and mocks for resolvers and controllers

+++ src/modules/post/post.resolver.ts
@@ -6,6 +6,8 @@ import { UpdatePostInput } from './dto/update-post.input';
 import { UseGuards } from '@nestjs/common';
 import { GqlAuthGuard } from '@auth/guards/gql-auth.guard';
 import { PaginationArgs } from '@dto/pagination.args';
+import { CurrentUser } from '@decorators/current-user.decorator';
+import { User } from '@prisma/client';
 
 @Resolver(() => Post)
 export class PostResolver {
@@ -13,8 +15,14 @@ export class PostResolver {
 
   @UseGuards(GqlAuthGuard)
   @Mutation(() => Post)
-  createPost(@Args('createPostInput') createPostInput: CreatePostInput) {
-    return this.postService.createOne(createPostInput);
+  createPost(
+    @Args('createPostInput') createPostInput: CreatePostInput,
+    @CurrentUser() user: User,
+  ) {
+    return this.postService.createOne({
+      ...createPostInput,
+      authorId: user.id,
+    });
   }
 
   @Query(() => [Post], { name: 'posts' })
@@ -29,10 +37,14 @@ export class PostResolver {
 
   @UseGuards(GqlAuthGuard)
   @Mutation(() => Post)
-  updatePost(@Args('updatePostInput') updatePostInput: UpdatePostInput) {
+  updatePost(
+    @Args('updatePostInput') updatePostInput: UpdatePostInput,
+    @CurrentUser() author: User,
+  ) {
     return this.postService.updateOne({
       where: {
         id: updatePostInput.id,
+        author,
       },
       data: {
         ...updatePostInput,
@@ -42,7 +54,10 @@ export class PostResolver {
 
   @UseGuards(GqlAuthGuard)
   @Mutation(() => Post)
-  removePost(@Args('id', { type: () => Int }) id: number) {
-    return this.postService.deleteOne({ id });
+  removePost(
+    @Args('id', { type: () => Int }) id: number,
+    @CurrentUser() author: User,
+  ) {
+    return this.postService.deleteOne({ id, author });
   }
 }
